Update to the latest data
This should fix the Air Pollutants and the Europe-EUplus issues. It also uses the _ in the file name for you to replace with a space.

I also added more climate results. These ones are yearly (instead of every 5 years) and have a longer time series. If the code can't handle yearly data, let me know.